Warnings fix
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@10647 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Private/src/TaskSchedulerImpl.cpp
@@ -1026,7 +1026,7 @@ void TaskThread::run() {
 
 bool TaskThread::event(QEvent *event) {
     QTimerEvent *timerEvent = NULL;
-    switch(event->type()) {
+    switch(static_cast<int>(event->type())) {
     case GET_NEW_SUBTASKS_EVENT_TYPE:
         getNewSubtasks();
         break;
